refactor: reorganize the structure of repo, #56
move the documents written in README.md into code comments
