+++ package.json
@@ -2,14 +2,15 @@
   "name": "serializr",
   "version": "1.1.12",
   "description": "Serialize and deserialize complex object graphs to JSON",
-  "main": "serializr.js",
+  "main": "lib/serializr.js",
+  "module": "lib/es/serializr.js",
   "typings": "serializr.d.ts",
   "scripts": {
     "test": "npm run build-test && node test/index",
-    "lint": "eslint serializr.js",
-    "prepublish": "npm run small-build && npm run build-docs",
-    "small-build": "uglifyjs -m sort,toplevel -c --screw-ie8 --preamble \"/** serializr - (c) Michel Weststrate 2016 - MIT Licensed */\" --source-map serializr.min.js.map -o serializr.min.js serializr.js",
-    "build-docs": "documentation readme serializr.js --github --section API",
+    "lint": "eslint src",
+    "prepublish": "npm run build && npm run build-docs",
+    "build": "node scripts/build.js",
+    "build-docs": "documentation readme src/serializr.js --github --section API",
     "build-test": "npm run build-test-babel && npm run build-test-ts",
     "build-test-ts": "tsc -p test/typescript",
     "build-test-babel": "babel test/babel/babel.js -o test/babel/babel-compiled.js",
@@ -32,21 +33,26 @@
     "url": "https://github.com/mobxjs/serializr/issues"
   },
   "files": [
-    "serializr.js",
-    "serializr.min.js",
-    "serializr.min.js.map",
+    "lib",
     "serializr.d.ts"
   ],
   "devDependencies": {
     "babel-cli": "^6.11.4",
+    "babel-eslint": "^7.2.3",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-stage-1": "^6.5.0",
     "coveralls": "^2.11.9",
     "documentation": "^4.0.0-beta9",
     "eslint": "^2.12.0",
+    "eslint-plugin-import": "^2.7.0",
     "istanbul": "^0.4.4",
+    "lodash.merge": "^4.6.0",
     "mobx": "^2.4.1 || ^3.0.0",
+    "rollup": "^0.49.2",
+    "rollup-plugin-commonjs": "^8.2.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-uglify": "^2.0.1",
     "tape": "^4.5.1",
     "typescript": "^2.1.4",
     "uglify-js": "^2.6.4"
